pyproject.toml
[project]
name = "lambda-otel-logging"
description = "Python logging formatter that looks like OpenTelemetry logs"
readme = "README.md"
authors = [{ name = "Bram Neijt", email = "bram@neijt.nl" }]
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]
dynamic = ["version"]
requires-python = ">=3.10"
dependencies = [
    "msgspec>=0.17.0",
    "opentelemetry-sdk>=1.20.0",
]

[dependency-groups]
dev = [
    "botocore>=1.42.4",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
]

[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.hatch.build.hooks.version]
path = "src/lambda_otel_logging/_version.py"

[tool.ruff.lint]
ignore = ["E501", "UP006", "FA100", "N803", "UP035", "UP007"]
extend-select = [
    "ANN",
    "B",
    "E",
    "FA",
    "I",
    "INP",
    "N",
    "Q",
    "RUF",
    "SIM",
    "UP",
]
